refactor(ui): extract useLogout hook shared between Sidebar and palette
